move clusterMetadata to cluster
