+++ common/cluster/utils.go
@@ -18,7 +18,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-package utils
+package cluster
 
 import "github.com/uber/cadence/common/persistence"
